+++ client/src/Types/Share.ts
@@ -1,5 +1,6 @@
-import { UserI } from "./User";
+
 import { RouteI } from "./Route";
+import { UserI } from "./User";
 
 export interface ShareI {
     _id?: string;
