--- server/Types/Share.ts
@@ -1,3 +1,0 @@
-import { UserI } from "./User.js";
-import { RouteI } from "./Route.js";
-
