+++ server/models/Share.ts
@@ -1,5 +1,14 @@
 import mongoose from "mongoose";
-import { ShareI } from "./UserModel";
+import { RouteI } from "./Route";
+import { UserI } from "./UserModel";
+
+export interface ShareI {
+  _id?: string;
+  owner: UserI;
+  route: RouteI;
+  password: string;
+  date: string;
+}
 
 export const shareSchema = new mongoose.Schema<ShareI>({
   owner: { type: mongoose.Schema.Types.ObjectId, ref: "User", required: true },
